backend/requirements.txt
google-generativeai
fastapi
uvicorn
python-multipart
pypdf
